Address ErrorProne immutable package warnings. (#196)

+++ src/test/java/tfw/immutable/ila/bitila/BitIlaCheck.java
@@ -15,7 +15,7 @@ public final class BitIlaCheck {
         assertThrows(IllegalArgumentException.class, () -> ila.get(array, -1, 0, 1));
         assertThrows(IllegalArgumentException.class, () -> ila.get(array, 0, -1, 1));
         assertThrows(IllegalArgumentException.class, () -> ila.get(array, 0, 0, -1));
-        assertThrows(IllegalArgumentException.class, () -> ila.get(array, array.length * Long.SIZE, 0, 1));
+        assertThrows(IllegalArgumentException.class, () -> ila.get(array, array.length * (long) Long.SIZE, 0, 1));
         assertThrows(IllegalArgumentException.class, () -> ila.get(array, 0, ilaLength * Long.SIZE, 1));
     }
 }
